khash: split up implementation
